Issue#22 - Added Test Cases
